requirements.txt
Django==2.0.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
psycopg2==2.7.4
pytz==2018.4
virtualenv==15.2.0
python-decouple==3.1
